split profile view into different component
